+++ src/Navigation/CurrentItemNavigation.tsx
@@ -5,7 +5,9 @@ import { Typography } from '@mui/material';
 import { DiscriminatedItem, ItemType } from '@graasp/sdk';
 
 import ItemMenu, { ItemMenuProps } from './ItemMenu';
-import { CenterAlignWrapper, ITEM_NAME_MAX_LENGTH, StyledLink } from './utils';
+import CenterAlignWrapper from './common/CenterAlignWrapper';
+import NavigationLink from './common/NavigationLink';
+import { ITEM_NAME_MAX_LENGTH } from './common/constants';
 
 export interface CurrentItemProps {
   item: DiscriminatedItem;
@@ -29,7 +31,7 @@ const CurrentItemNavigation = ({
 }: CurrentItemProps): JSX.Element | null => {
   return (
     <CenterAlignWrapper>
-      <StyledLink
+      <NavigationLink
         id={buildBreadcrumbsItemLinkId?.(item.id)}
         key={item.id}
         to={buildToItemPath(item?.id)}
@@ -37,7 +39,7 @@ const CurrentItemNavigation = ({
         <Typography>
           {truncate(item.name, { length: ITEM_NAME_MAX_LENGTH })}
         </Typography>
-      </StyledLink>
+      </NavigationLink>
       {(item.type === ItemType.FOLDER || showArrow) && (
         <ItemMenu
           useChildren={useChildren}
