+++ src/Navigation/Navigation.tsx
@@ -8,12 +8,7 @@ import CurrentItemNavigation from './CurrentItemNavigation';
 import ExtraItemsNavigation, { ExtraItem } from './ExtraItemsNavigation';
 import { ItemMenuProps } from './ItemMenu';
 import ParentsNavigation from './ParentsNavigation';
-
-const CenterAlignWrapper = styled('div')({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
+import CenterAlignWrapper from './common/CenterAlignWrapper';
 
 const StyledBreadcrumbs = styled(Breadcrumbs)(({ theme }) => ({
   '& ol': {
