+++ src/Navigation/common/CenterAlignWrapper.tsx
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material';
+
+const CenterAlignWrapper = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+});
+export default CenterAlignWrapper;
