+++ src/Navigation/common/NavigationLink.tsx
@@ -1,0 +1,9 @@
+import { styled } from '@mui/material';
+
+import { Link } from 'react-router-dom';
+
+const NavigationLink = styled(Link)({
+  textDecoration: 'none',
+  color: 'text.primary',
+});
+export default NavigationLink;
