+++ src/Navigation/common/constants.ts
@@ -1,0 +1,1 @@
+export const ITEM_NAME_MAX_LENGTH = 15;
